upd figures and furmulas
